requirements.txt
selenium==4.21.0
webdriver-manager==4.0.1
pytest==8.2.0
pytest-bdd==7.2.0
pytest-xdist==3.5.0
pytest-html==4.1.1
requests==2.31.0
jsonschema==4.22.0
pydantic==2.7.1
Appium-Python-Client==3.1.1
SQLAlchemy==2.0.30
pymysql==1.1.0
anthropic==0.28.0
mcp==1.3.0
openpyxl==3.1.2
pandas==2.2.2
faker==25.0.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
python-dotenv==1.0.1
PyYAML==6.0.1
loguru==0.7.2
tenacity==8.3.0
Pillow==10.3.0